RideManager
Conecting both contracts with a manager , this helps to emulate better the transaction , it stores the info of both parts driver and customer.
Troubleshoot, the blocks are succesfully updating after been called by driver or customer, the data is wrong so try to fix.
then if that is solved  the only  obligatory parameter is the rideManager as from there the addres can be extracted.