Ride manager fixed
manager instance has been successfully fixed for interaction, generating valuable information for issue resolution. Further development could include implementing timestamps for added security & distance agreement or something like that

Key functionalities, such as starting and ending a ride, recording the time when the driver accepted the ride, and managing the distance, can be enhanced for mutual agreement between the driver and the customer.

While both the taxi and the customer currently have access to all parts of the contract, it's advisable to maintain separation for better practices. This separation may not be as significant in a phone or web application where user interfaces can streamline the process.